+++ resources/views/livewire/translation-editor-modal-empty.blade.php
@@ -1,1 +1,1 @@
-{{-- Empty view when translation editor is disabled --}}
+<div></div>