+++ resources/views/livewire/translation-editor-modal-wrapper.blade.php
@@ -1,16 +1,9 @@
-<!-- Translation Editor Modal Wrapper -->
 @if (auth()->check())
     @php
-        $user = auth()->user();
         $toggleColumn = config('translation-editor.auth.toggle_column', 'translation_modal_enabled');
-        $isEnabled = $user->{$toggleColumn} ?? false;
     @endphp
 
-    @if ($isEnabled)
+    @if (auth()->user()->{$toggleColumn} ?? false)
         @livewire('translation-editor-modal')
-    @else
-        <!-- Translation editor disabled for user. Toggle column: {{ $toggleColumn }}, Value: {{ var_export($user->{$toggleColumn} ?? null, true) }} -->
     @endif
-@else
-    <!-- User not authenticated -->
 @endif